chore: refactor makefile structure
chage the relation between root makefile and the makefile of each pakcage.
after updating, the `dev/base.makefile` will be the base makefile,
and the `packages/*/makefile` extends the base one.
the targets in base makefile can be override in the makefile of package (need to remove the postfix `-default`)