feat(react-article-components): add lazyload to <img>
This change adds isLazyLoaded as a prop to img-with-placeholder to
add `loading="lazy"` to <img> which lazyloads the image.

Also, set the leading image of articles not to be lazyloaded by
default because it relates to (LCP) largest contentful paint metric.
Usually, the leading image is the LCP image and should not be
lazy-loaded.