feat(index-page): add isLazyLoaded prop to img-wrapper
This patch adds isLazyLoaded prop to imag-wrapper to set the attribute
`loading="lazy"` by default.

Also, set the image in editor-picks not to be lazyloaded by
default because it relates to (LCP) largest contentful paint metric.
Usually, the image in editor-picks on desktop is the LCP image and
should not be lazy-loaded.